removed copy folder contents remove contents of copy
everything in copy is synced from the private dotfiles folder. Both directories serve a similar purpose, and I hoped to keep copy around with _some_ reference settings and git config items, but every time dotfiles runs the copy step backs up my private dotfiles, and then my private dotfiles back up the copy. It's a strange and unfruitful dance that clutters up the backup folder unecessarily. It's not much space, but over the course of a year or two it may grow substantially with copies of copies of copies.