initial zsh support
I intend to switch to zsh from bash. It's the default on new MacOS installs not and while it's possible to switch back to bash its a pain. There is no single pref in MacOS that actually switches things back over all contexts, editor terminals, etc. Easier to migrate now.

I'm sure this is only the start of what's necessary to work with zsh