materialize the full starship config
this does not include keys with unset values, but does include previous custom values and current defaults at the time of commit

This was materialized with `starship print-config` and should help with understanding some of the undesired default `format` text.

Follow on commits may remove some of these environments. It's the starship default, but I don't use, recommend, or recognize a lot of these tools and formats